backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=21.2.0
azure-cosmos>=4.5.0
python-dotenv>=1.0.0
